Add files via upload
As the predictive analytic concept was increasing its popularity, my team managed to come up a model to obtain the best performance on predicting the odd of whether or not NFL player being drafted. This project involved in the usage of Rstudio date from its beginning. Models such as decision tree, logistic regression and random forests are compared for prediction power. While the model performance varied, it is still preferable for teams to consider logistic regression for prediction purpose.