Add files via upload
The season of basketball is around the corner, my team have things organized for the matter in hand with a model to predict the odd of whether or not a shot is made by a NBA player using feature selection. This project involved in the usage of python, pyspark, sqlite3 and google_colab. As contributing equally to different sections of coding in colab, my team made the best use of data retrieval and feature selection with a platform above to connect the classification model with light machine learning techniques. As a result, it is as clear as the day that the shot distance of a player explains the odd of the shot attempt the most.