Add files via upload
As the statistical methods were raising attention in data analytical field, I put a semester length of class contents into practice using Rstudio. The goal of this project is to build a price predicting model using housing geospatial locations with the individual unique properties (e.g. year of transaction) of each house in terms of its surroundings. Along with this focus, I also predicted the geospatial location of the metro station by using visualization techniques. 
PS.The R code template is attached to the end of the document, please feel free to borrow them if they may be beneficial to you.